requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
crispy-tailwind==1.0.3
Django==5.2.5
django-crispy-forms==2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
mysqlclient==2.2.7
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
stripe==12.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
